+++ archlens-in-vscode/src/views/FileTreeView.ts
